Refactor tree2 snapshot tests to support multiple trees (#16370)
## Description

Structures the snapshot testing in tree2 to more easily support trees
with different data/shapes. This will let us more easily get better
coverage for snapshot-impacting features.
